requirements.txt
google-adk>=1.0.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
litellm>=1.0.0,<2.0.0
pytest>=8.0.0,<9.0.0
